modify grad_handle func and move common func to tensor_utils.py
