+++ colossalai/gemini/memory_tracer/param_tracer_wrapper.py
@@ -1,9 +1,8 @@
 import torch.nn
 
-from colossalai.tensor.colo_parameter import ColoParameter
 from colossalai.tensor.param_op_hook import ParamOpHookManager
 from colossalai.gemini.ophooks.param_trace_hook import ParamTracerHook
-from colossalai.gemini.chunk.chunk import free_storage
+from colossalai.gemini.tensor_utils import free_storage
 from colossalai.nn.parallel.data_parallel import _cast_float
 from functools import partial
 
@@ -19,17 +18,16 @@ class ParamTracerWrapper():
         self.param_op_hook = ParamTracerHook(dtype)
 
         for p in module.parameters():
-            # assert isinstance(p, ColoParameter)
             p.data = p.data.to(dtype)
             if p.requires_grad:
-                p.register_hook(partial(self.grad_handle, p))
+                p.register_hook(partial(self.grad_handle))
 
         self._cast_buffers_to_cuda_dtype()
 
     def __call__(self, *args, **kwargs):
         return self.forward(*args, **kwargs)
 
-    def grad_handle(self, p, grad):
+    def grad_handle(self, grad):
         free_storage(grad)
 
     def _pre_forward(self):
